Static JFR events
